Added a crouch
Added crouch to my basic movement. It stutters a bit when returning from crouch to standing, so have to figure that out. I believe it has to do with the fact that the collider is intersecting with the ground. Also added a bunch of folders to the git ignore list.